clean up mess
