--- inventory_manager.py
@@ -1,27 +1,0 @@
-from inventory import *
-from database import *
-
-
-class InventoryManager:
-    def __init__(self):
-        pass
-
-    def search_product(self, query):
-        """Search for product in inventory"""
-        pass
-
-    def add_product(self, product):
-        """add new product to inventory"""
-        pass
-
-    def remove_product(self, product_name):
-        """Remove existing product from inventory"""
-        pass
-
-    def update_quantity(self, product_name, quantity):
-        """Update quantity of a product in inventory"""
-        pass
-
-    def get_total_inventory_value(self):
-        """Calculate total value of the entire inventory"""
-        pass
